fix: <Backlog> Ajustes

+++ api/src/main/java/com/khali/api3/repositories/NotificationRepository.java
@@ -7,6 +7,6 @@ import com.khali.api3.domain.notification.Notification;
 
 // manipulação das tabelas
 @RepositoryRestResource
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
 
 }
